refactor: rename class - LottoNumbers->Lotto - LottoTicket->Lottos
