--- src/main/java/lotto/domain/LottoTicket.java
@@ -1,17 +1,0 @@
-package lotto.domain;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class LottoTicket {
-    private final List<Lotto> lottos;
-
-    public LottoTicket(List<Lotto> lottos) {
-        this.lottos = new ArrayList<>(lottos);
-    }
-
-    public List<Lotto> values() {
-        return Collections.unmodifiableList(lottos);
-    }
-}
