+++ src/main/java/lotto/domain/Lottos.java
@@ -1,0 +1,15 @@
+package lotto.domain;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class Lottos {
+    private final List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos) {
+        this.lottos = Collections.unmodifiableList(new ArrayList<>(lottos));
+    }
+}
