+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -6,6 +6,6 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class WinningNumbers {
-    private final LottoNumbers numbers;
+    private final Lotto numbers;
     private final int bonusNumber;
 }
